requirements.txt
click==8.0.3
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap4==4.0.2
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
SQLAlchemy==1.4.28
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.0
